requirements.txt
Flask==2.3.3
Flask_SQLAlchemy==3.1.1
python-dotenv==1.0.1
Werkzeug==3.0.1
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.3.0
flake8==7.0.0
huggingface_hub==0.23.0
